Add to_constraint mapping for cpu_workload.

+++ app/models/ram_workload.rb
@@ -12,7 +12,7 @@ class RamWorkload < ActiveRecord::Base
   end
 
   def to_constraint
-    self.ingredient.ram_constraint.destroy
+    self.ingredient.ram_constraint.destroy if self.ingredient.ram_constraint.present?
     self.ingredient.ram_constraint = RamConstraint.create(
       min_ram: self.ram_mb_required + num_simultaneous_users * ram_mb_growth_per_user
     )
